Implemented slider bounce mechanic
The ball will behave differently when it bounces off the slider - compared to a bounce off on a wall. It will keep it's angle when hitting the wall - but get a new angle when hitting the slider - depending on the region of the slider it hits. The current implementation is a rough estimation of the angles used by the original Breakout.